+++ test/specs/components/desktop/prevent-double-click.component.spec.jsx
@@ -3,24 +3,68 @@ var reactTestUtils = React.addons.TestUtils;
 var PreventDoubleClick = require('../../../../web/app/components/desktop/prevent-double-click.component.jsx');
 var testHelper = require('../../../test-helper');
 var _ = require('lodash');
+var sinon = require('sinon');
+var bluebird = require('bluebird');
+var userStore = require('../../../../web/app/stores/user.store');
+
+var getPreventDoubleClickPageComponent = function(mainElement) {
+  return reactTestUtils.findRenderedComponentWithType(mainElement, PreventDoubleClick);
+};
 
 describe('prevent double click page', function() {
+  before(function() {
+    sinon.stub(userStore, 'getUser', function(userId) {
+      var defer = bluebird.defer();
+
+      if (userId === 124) {
+        defer.resolve(testHelper.mockedData.users['124']);
+      }
+
+      return defer.promise;
+    });
+  });
+
+  after(function() {
+    userStore.getUser.restore();
+  });
+
   beforeEach(function() {
     testHelper.resetStoresCachedData('Menu', 'Application');
   });
 
+  it('should have correct menu', function(done) {
+    testHelper.testPage('/prevent-double-click', function(mainElement) {
+      testHelper.testMenu(mainElement, [{
+        href: 'desktop',
+        display: 'Desktop',
+        className: 'header-desktop-link'
+      }, {
+        href: 'prevent-double-click',
+        display: 'Prevent Double Click',
+        className: 'header-prevent-double-click-link'
+      }, {
+        href: 'with-resolves',
+        display: 'With Resolves',
+        className: 'header-with-resolves-link'
+      }]);
+      done();
+    });
+  });
+
   it('should have h1', function(done) {
-    testHelper.getRouterComponent(PreventDoubleClick, '/prevent-double-click', this, function() {
-      var h1 = reactTestUtils.findRenderedDOMComponentWithTag(this.component, 'h1');
+    testHelper.testPage('/prevent-double-click', function(mainElement) {
+      var preventDoubleClickPageComponent = getPreventDoubleClickPageComponent(mainElement);
+      var h1 = reactTestUtils.findRenderedDOMComponentWithTag(preventDoubleClickPageComponent, 'h1');
 
       expect(h1).to.be.defined;
       expect(h1.props.children).to.equal('Prevent Double Click');
       done();
-    }.bind(this));
+    });
   });
 
   it('should rendered the 2 buttons', function(done) {
-    testHelper.getRouterComponent(PreventDoubleClick, '/prevent-double-click', this, function() {
+    testHelper.testPage('/prevent-double-click', function(mainElement) {
+      var preventDoubleClickPageComponent = getPreventDoubleClickPageComponent(mainElement);
       expectedData = [{
         disabled: false,
         children: 'test'
@@ -29,7 +73,7 @@ describe('prevent double click page', function() {
         children: 'Prevent Double Click Other Buttons'
       }];
 
-      var buttons = reactTestUtils.scryRenderedDOMComponentsWithTag(this.component, 'button');
+      var buttons = reactTestUtils.scryRenderedDOMComponentsWithTag(preventDoubleClickPageComponent, 'button');
 
       expect(buttons).to.be.defined;
       expect(buttons.length).to.equal(expectedData.length);
@@ -46,29 +90,78 @@ describe('prevent double click page', function() {
         });
       });
       done();
-    }.bind(this));
+    });
   });
 
   it('should disable first button when clicking on second button', function(done) {
-    testHelper.getRouterComponent(PreventDoubleClick, '/prevent-double-click', this, function() {
-      var buttons = reactTestUtils.scryRenderedDOMComponentsWithTag(this.component, 'button');
+    testHelper.testPage('/prevent-double-click', function(mainElement) {
+      var preventDoubleClickPageComponent = getPreventDoubleClickPageComponent(mainElement);
+      var buttons = reactTestUtils.scryRenderedDOMComponentsWithTag(preventDoubleClickPageComponent, 'button');
 
       reactTestUtils.Simulate.click(buttons[1]);
 
       expect(buttons[0].props.disabled).to.be.true;
       done();
-    }.bind(this));
+    });
   });
 
   it('should enable first button when when it is already disabled when clicking on second button', function(done) {
-    testHelper.getRouterComponent(PreventDoubleClick, '/prevent-double-click', this, function() {
-      var buttons = reactTestUtils.scryRenderedDOMComponentsWithTag(this.component, 'button');
+    testHelper.testPage('/prevent-double-click', function(mainElement) {
+      var preventDoubleClickPageComponent = getPreventDoubleClickPageComponent(mainElement);
+      var buttons = reactTestUtils.scryRenderedDOMComponentsWithTag(preventDoubleClickPageComponent, 'button');
 
       reactTestUtils.Simulate.click(buttons[1]);
       reactTestUtils.Simulate.click(buttons[1]);
 
       expect(buttons[0].props.disabled).to.be.false;
       done();
-    }.bind(this));
+    });
+  });
+
+  it('should have link to prevent double click page', function(done) {
+    testHelper.testPage('/prevent-double-click', function(mainElement) {
+      var link = reactTestUtils.scryRenderedDOMComponentsWithClass(mainElement, 'header-prevent-double-click-link');
+
+      expect(link.length).to.equal(1);
+      done();
+    });
+  });
+
+  it('should have link to desktop page', function(done) {
+    var steps = [];
+
+    steps.push(function(mainElement) {
+      var link = reactTestUtils.findRenderedDOMComponentWithClass(mainElement, 'header-desktop-link');
+
+      testHelper.simulateRouterLinkClick(link);
+    });
+
+    steps.push(function(mainElement) {
+      var page = reactTestUtils.scryRenderedDOMComponentsWithClass(mainElement, 'p-desktop');
+
+      expect(page.length).to.equal(1);
+      done();
+    });
+
+    testHelper.testPage('/prevent-double-click', steps);
+  });
+
+  it('should have link to with resolves page', function(done) {
+    var steps = [];
+
+    steps.push(function(mainElement) {
+      var link = reactTestUtils.findRenderedDOMComponentWithClass(mainElement, 'header-with-resolves-link');
+
+      testHelper.simulateRouterLinkClick(link);
+    });
+
+    steps.push(function(mainElement) {
+      var userId = reactTestUtils.findRenderedDOMComponentWithClass(mainElement, 'user-id');
+
+      expect(userId.props.children).to.equal(124);
+      done();
+    });
+
+    testHelper.testPage('/prevent-double-click', steps);
   });
 });