+++ test/specs/stores/menu.store.spec.js
@@ -10,21 +10,26 @@ describe('menu data', function() {
     expect(Object.keys(store._cachedData).length).to.equal(2);
     expect(store._cachedData.menus).to.deep.equal({
       desktop: [{
-        href: '/desktop',
-        display: 'Desktop'
+        href: 'desktop',
+        display: 'Desktop',
+        className: 'header-desktop-link'
       }, {
-        href: '/prevent-double-click',
-        display: 'Prevent Double Click'
+        href: 'prevent-double-click',
+        display: 'Prevent Double Click',
+        className: 'header-prevent-double-click-link'
       }],
       preventDoubleClick: [{
-        href: '/desktop',
-        display: 'Desktop'
+        href: 'desktop',
+        display: 'Desktop',
+        className: 'header-desktop-link'
       }, {
-        href: '/prevent-double-click',
-        display: 'Prevent Double Click'
+        href: 'prevent-double-click',
+        display: 'Prevent Double Click',
+        className: 'header-prevent-double-click-link'
       }, {
-        href: '/with-resolves',
-        display: 'With Resolves'
+        href: 'with-resolves',
+        display: 'With Resolves',
+        className: 'header-with-resolves-link'
       }]
     });
     expect(store._cachedData.activeMenu).to.equal('desktop');
@@ -40,11 +45,13 @@ describe('menu data', function() {
 
   it('should be able to get the active menu set', function() {
     expect(store.getMenu()).to.deep.equal([{
-      href: '/desktop',
-      display: 'Desktop'
+      href: 'desktop',
+      display: 'Desktop',
+        className: 'header-desktop-link'
     }, {
-      href: '/prevent-double-click',
-      display: 'Prevent Double Click'
+      href: 'prevent-double-click',
+      display: 'Prevent Double Click',
+        className: 'header-prevent-double-click-link'
     }]);
   });
 
@@ -54,14 +61,17 @@ describe('menu data', function() {
     });
 
     expect(store.getMenu()).to.deep.equal([{
-      href: '/desktop',
-      display: 'Desktop'
+      href: 'desktop',
+      display: 'Desktop',
+        className: 'header-desktop-link'
     }, {
-      href: '/prevent-double-click',
-      display: 'Prevent Double Click'
+      href: 'prevent-double-click',
+      display: 'Prevent Double Click',
+        className: 'header-prevent-double-click-link'
     }, {
-      href: '/with-resolves',
-      display: 'With Resolves'
+      href: 'with-resolves',
+      display: 'With Resolves',
+        className: 'header-with-resolves-link'
     }]);
   });
 });