+++ test/test-helper.js
@@ -6,12 +6,14 @@ var storeLocations = {
 };
 var initialStoresCachedData = {};
 var Router = require('react-router');
+var Link = Router.Link;
 var Route = Router.Route;
 var React = require('react/addons');
 var reactTestUtils = React.addons.TestUtils;
 var TestLocation = require('react-router/lib/locations/TestLocation');
 var fibers = require('fibers');
 var mockedData = require('../web/mocked-api/data/index');
+var Header = require('../web/app/components/core/header.component.jsx');
 
 //store the original state of all the stores
 _.forEach(storeLocations, function(path, storeName) {
@@ -31,22 +33,31 @@ module.exports = {
     });
   },
 
-  getRouterComponent: function(Component, path, attachComponentTo, callback) {
+  testPage: function(initialPath, steps) {
+    if (!_.isArray(steps)) {
+      steps = [steps];
+    };
+
     var component;
+    var routerMainComponent;
     var div = document.createElement('div');
     var routes = require('../web/app/components/core/routes.jsx');
-    var location = new TestLocation([path]);
+    var location = new TestLocation([initialPath]);
 
     Router.run(routes, location, function (Handler) {
-      var mainComponent = React.render(React.createFactory(Handler)({}), div);
-      attachComponentTo.component = reactTestUtils.findRenderedComponentWithType(mainComponent, Component);
-      callback();
+      var step = steps.shift();
+
+      //TODO: research: not sure why or if I need this here (https://github.com/rackt/react-router/issues/991)
+      this.unmountComponent(routerMainComponent);
+
+      routerMainComponent = React.render(React.createFactory(Handler)({}), div);
+      step(routerMainComponent);
     }.bind(this));
   },
 
   unmountComponent: function(component) {
     if(component && component.isMounted()) {
-      console.log(component.constructor.displayName);
+      //console.log('unmounted: ' + component.constructor.displayName);
       React.unmountComponentAtNode(component.getDOMNode().parentNode);
     }
   },
@@ -118,5 +129,30 @@ module.exports = {
     component.type.prototype.__reactAutoBindMap[eventHandlerName].restore();
   },
 
-  mockedData: mockedData
+  mockedData: mockedData,
+
+  simulateRouterLinkClick: function(linkComponent) {
+    reactTestUtils.Simulate.click(linkComponent, {button: 0});
+  },
+
+  //TODO: remove: this is specific to the menu setup that is in place so either remove this method (if you are using a different menu system) or remove this
+  //TODO: remove: comment if you are using the provided menu system
+  testMenu: function(mainElement, expectedMenuData) {
+    var headerComponent = reactTestUtils.findRenderedComponentWithType(mainElement, Header);
+    var ulElement = reactTestUtils.findRenderedDOMComponentWithTag(headerComponent, 'ul');
+
+    expect(ulElement).to.be.defined;
+
+    liElements = reactTestUtils.scryRenderedDOMComponentsWithTag(ulElement, 'li');
+
+    expect(liElements.length).to.equal(expectedMenuData.length);
+
+    _.forEach(liElements, function(liElement, key) {
+      var linkProps = reactTestUtils.findRenderedComponentWithType(liElement, Link).props;
+
+      expect(linkProps.to).to.equal(expectedMenuData[key].href);
+      expect(linkProps.children).to.equal(expectedMenuData[key].display);
+      expect(linkProps.className).to.equal(expectedMenuData[key].className);
+    });
+  }
 };