+++ web/app/components/core/application.component.jsx
@@ -5,15 +5,17 @@ var RouteHandler = require('react-router').RouteHandler;
 //TODO: remove: for testing assets rewrit eprocess
 var removeMe = '/app/components/core/remove-me.svg#icon-small';
 
-var Application = React.createClass({
-  render: function applicationComponentRender() {
-    return (
-      <div className="application">
-        <Header />
-        <RouteHandler />
-      </div>
-    );
-  }
-});
+var application = {};
 
-module.exports = Application;
+application.displayName = 'Application';
+
+application.render = function applicationRender() {
+  return (
+    <div className="application">
+      <Header />
+      <RouteHandler />
+    </div>
+  );
+};
+
+module.exports = React.createClass(application);