+++ web/app/components/core/header.component.jsx
@@ -2,40 +2,44 @@ var React = require('react/addons');
 var menuStore = require('../../stores/menu.store');
 var Link = require('react-router').Link;
 
-var Header = React.createClass({
-  getInitialState: function headerComponentGetInitialState() {
-    return {
-      menu: menuStore.getMenu()
-    };
-  },
+var header = {};
 
-  onChange: function headerComponentOnChange() {
-    this.setState({
-      menu: menuStore.getMenu()
-    });
-  },
+header.displayName = 'Header';
 
-  componentDidMount: function headerComponentComponentDidMount() {
-    menuStore.on('activeMenuUpdated', this.onChange);
-  },
+header.getInitialState = function headerComponentGetInitialState() {
+  return {
+    menu: menuStore.getMenu()
+  };
+};
 
-  componentWillUnmount: function headerComponentComponentWillUnmount() {
-    menuStore.removeListener('activeMenuUpdated', this.onChange);
-  },
+header.onChange = function headerComponentOnChange() {
+  this.setState({
+    menu: menuStore.getMenu()
+  });
+};
 
-  render: function headerComponentRender() {
-    return (
-      <header>
-        <ul>
-          {this.state.menu.map(function headerComponentMapMenu(menuItem) {
-            return (<li key={menuItem.href}>
-              <Link to={menuItem.href}>{menuItem.display}</Link>
-            </li>);
-          })}
-        </ul>
-      </header>
-    );
-  }
-});
+header.componentDidMount = function headerComponentComponentDidMount() {
+  menuStore.on('activeMenuUpdated', this.onChange);
+};
 
-module.exports = Header;
+header.componentWillUnmount = function headerComponentComponentWillUnmount() {
+  menuStore.removeListener('activeMenuUpdated', this.onChange);
+};
+
+header.render = function headerComponentRender() {
+  return (
+    <header>
+      <ul>
+        {this.state.menu.map(function headerComponentMapMenu(menuItem) {
+          return (
+            <li key={menuItem.href}>
+              <Link className={menuItem.className} to={menuItem.href}>{menuItem.display}</Link>
+            </li>
+          );
+        })}
+      </ul>
+    </header>
+  );
+};
+
+module.exports = React.createClass(header);