+++ web/app/components/desktop/module.jsx
@@ -8,8 +8,8 @@ var WithResolves = require('./with-resolves.component.jsx');
 
 module.exports = {
   routes: [
-    <Route key="1" path="/desktop" handler={Desktop} />,
-    <Route key="2" path="/prevent-double-click" handler={PreventDoubleClick} />,
-    <Route key="3" path="/with-resolves" handler={WithResolves} />
+    <Route key="1" name="desktop" path="/desktop" handler={Desktop} />,
+    <Route key="2" name="prevent-double-click" path="/prevent-double-click" handler={PreventDoubleClick} />,
+    <Route key="3" name="with-resolves" path="/with-resolves" handler={WithResolves} />
   ]
-}
+};