+++ web/app/components/desktop/with-resolves.component.jsx
@@ -1,57 +1,58 @@
 var React = require('react/addons');
 var menuStore = require('../../stores/menu.store');
-var bluebird = require('bluebird');
 var userStore = require('../../stores/user.store');
 
 var preLoadedData = {};
+var withResolves = {};
 
-var WithResolves = React.createClass({
-  statics: {
-    willTransitionTo: function withResolvesComponentWillTransitionTo(transition, params, query, callback) {
-      userStore.getUser(124).then(function(user) {
-        preLoadedData.user = user;
-        callback();
-      });
-    }
-  },
+withResolves.displayName = 'WithResolves';
 
-  getInitialState: function desktopComponentGetInitialState() {
-    return {
-      user: preLoadedData.user
-    };
-  },
-
-  componentDidMount: function withResolvesComponentComponentDidMount() {
-    menuStore.update({
-      menuName: 'desktop'
+withResolves.statics = {
+  willTransitionTo: function withResolvesComponentWillTransitionTo(transition, params, query, callback) {
+    userStore.getUser(124).then(function withResolvesComponentWillTransitionToSuccess(user) {
+      preLoadedData.user = user;
+      callback();
     });
-  },
+  }
+};
 
-  renderUserData: function desktopComponentRenderUserData() {
-    var userData = null;
+withResolves.getInitialState = function desktopComponentGetInitialState() {
+  return {
+    user: preLoadedData.user
+  };
+};
 
-    if (this.state.user) {
-      userData = (
-        <ul>
-          <li className="user-id">{this.state.user.id}</li>
-          <li className="user-username">{this.state.user.username}</li>
-          <li className="user-first-name">{this.state.user.firstName}</li>
-          <li className="user-last-name">{this.state.user.lastName}</li>
-        </ul>
-      );
-    }
+withResolves.componentDidMount = function withResolvesComponentComponentDidMount() {
+  menuStore.update({
+    menuName: 'desktop'
+  });
+};
 
-    return userData;
-  },
+withResolves.renderUserData = function desktopComponentRenderUserData() {
+  var userData = null;
 
-  render: function withResolvesComponentRender() {
-    return (
-      <span>
-        <h1 id="test" className="test">{window.i18n['desktop/with-resolves'].header()}</h1>
-        {this.renderUserData()}
-      </span>
+  /*istanbul ignore else*/
+  if (this.state.user) {
+    userData = (
+      <ul>
+        <li className="user-id">{this.state.user.id}</li>
+        <li className="user-username">{this.state.user.username}</li>
+        <li className="user-first-name">{this.state.user.firstName}</li>
+        <li className="user-last-name">{this.state.user.lastName}</li>
+      </ul>
     );
   }
-});
 
-module.exports = WithResolves;
+  return userData;
+};
+
+withResolves.render = function withResolvesComponentRender() {
+  return (
+    <span className="p-with-resolves">
+      <h1 id="test" className="test">{window.i18n['desktop/with-resolves'].header()}</h1>
+      {this.renderUserData()}
+    </span>
+  );
+};
+
+module.exports = React.createClass(withResolves);