+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # master
 
+- better support for testing react router based components
 - assets rewrite add `/build/` to path with configurable option to add `/static/[SHA]/`
 - added support for copying directory in copy static assets task
 - added support for #... in assets rewrite regex